% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{count_by256}
\alias{count_by256}
\title{Aggregating helpers}
\usage{
count_by256(DT, by = NULL, count_col = "N")
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{by}{(string) A column of \code{DT}, the count of which is desired.}

\item{count_col}{(string) The name of the column in the result containing
the counts.}
}
\value{
For:
\describe{
\item{\code{count_by256}}{A tally of \code{by}.}
}
}
\description{
Aggregating helpers
}
