% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor256.R
\name{factor256}
\alias{factor256}
\alias{recompose256}
\alias{levels.factor256}
\alias{is.factor256}
\alias{factor256_in}
\alias{factor256_notin}
\title{Factors of fewer than 256 elements}
\usage{
factor256(x, levels = NULL)

recompose256(f)

\method{levels}{factor256}(x)

is.factor256(x)

factor256_in(x, tbl)

factor256_notin(x, tbl)
}
\arguments{
\item{x}{An atomic vector with fewer than 256 elements.}

\item{levels}{An optional vector of the unique values of \code{x}.}

\item{f}{A raw vector of class \code{factor256}.}

\item{tbl}{The table of values to lookup in \code{f}.}
}
\value{
A raw vector of class \code{factor256}.

Values in \code{x} absent from \code{levels} are mapped to \code{00}.

\code{recompose256} is the inverse operation.
}
\description{
Whereas base R's factors as based on 32-bit integer vectors,
\code{factor256} uses 8-bit raw vectors to minimize memory footprints
}
