% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlace_raws.R
\name{interlace256}
\alias{interlace256}
\alias{deinterlace256}
\title{Interlace raw vectors}
\usage{
interlace256(w, x, y = NULL, z = NULL)

deinterlace256(u)
}
\arguments{
\item{w, x, y, z}{Raw vectors. A vector may be \code{NULL} if fewer than four
vectors need to be compressed.}

\item{u}{An integer vector.}
}
\value{
\code{interlace256} Return an integer vector, compressing raw vectors.
\code{deinterlace256} is the inverse operation, returning a list of four raw vectors.
}
\description{
Some processes do not accept raw vectors so it can be
necessary to convert our vectors to integers.
}
